% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cap.gen.joined}
\alias{cap.gen.joined}
\title{Joined Electrical Capacity and Generation Output}
\format{A data framewith 127044 and 6 variables}
\usage{
data(cap.gen.joined)
}
\description{
This is the dataset that reports capacity and generation using the constructed fuel and technological
categorical variables, which describe the data more broadly than the native fuel and primemover codes.
}
\details{
These two datasets are joined by two methods. First, in 2001-2002, we join by yr, plntcode, and fuel. We
do this because the Generation dataset reports primemovers exclusively as NA for these years. In all other years,
we join by yr, plntcode, primemove, and fuel. Neither method uses utilcode, as there are significant
inconsistencies within both the Generation and Capacity datasets, which prevent adequate joining between
the two datasets.

We take the inner_join() between these datasets using the keys indicated above. Finally, that join is mapped
to the overnightcategory and fuel variables contained in our mapping file. Mapping to these new keys
requires aggregation, as multiple pm-f codes map to the same overnightcategory and fuel pair. The resulting
dataset is used in the calculation of capacity factor.

\describe {
\item{yr}{Year of reported data}
\item{plntcode}{EIA-assigned plant code}
\item{overnightcategory}{Technology category}
\item{fuel.general}{Fuel category}
\item{nameplate}{Reported nameplate capacity, MW}
\item{generation}{Reported net electrical output, MWh}
}
}
