% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{capacityfactors.clamp}
\alias{capacityfactors.clamp}
\title{Capacity Factors, clamped to 1 from above}
\format{A data framewith 101437 and 6 variables}
\usage{
data(capacityfactors.clamp)
}
\description{
Plant-specific capacity factors for specified fuel and technology choices.
}
\details{
This calculation can be found in data-raw/costs/capacityfactors.R. Observations with a
calculated capacity factor greater than one are set equal to one.

\describe{
\item{yr}{Year of reported data}
\item{utilcode}{EIA-assigned utility code}
\item{plntcode}{EIA-assigned plant code}
\item{overnightcategory}{Technology category}
\item{fuel.general}{Fuel category}
\item{capacityfactor}{8760*capacity/generation, unitless}
}
}
