% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{capacityfactors}
\alias{capacityfactors}
\title{Capacity Factors}
\format{A data framewith 101437 and 6 variables}
\usage{
data(capacityfactors)
}
\description{
Plant-specific capacity factors for specified fuel and technology choices.
}
\details{
This calculation can be found in data-raw/costs/capacityfactors.R. Due to data errors,
some capacity factors are greater than one.

\describe{
\item{yr}{Year of reported data}
\item{utilcode}{EIA-assigned utility code}
\item{plntcode}{EIA-assigned plant code}
\item{overnightcategory}{Technology category}
\item{fuel.general}{Fuel category}
\item{capacityfactor}{8760*capacity/generation, unitless}
}
}
