% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{capacityfactors.sup}
\alias{capacityfactors.sup}
\title{Capacity Factors, Supplemental}
\format{A data framewith 15 and 2 variables}
\usage{
data(capacityfactors.sup)
}
\description{
Technology-specific capacity factors taken from the EIA Electricity Power Monthly.
}
\details{
\describe{
\item{overnightcategory}{Technology category}
\item{capacityfactor}{8760*capacity/generation, unitless}
}
}
