% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{capitalcosts}
\alias{capitalcosts}
\title{Capital Costs}
\format{A data framewith 225 and 5 variables}
\source{
\url{https://www.eia.gov/outlooks/aeo/archive.php}
}
\usage{
data(capitalcosts)
}
\description{
Input assumptions to EIA's National Energy Model System, which is used to produce the Annual
Energy Outlook report, for 1996-2015.
}
\details{
Tables from each report appear under the name 'Cost and Performance Characteristics of New
Central Station Electricity Generating Technologies' and were manually compiled into one table.
This file is an input into the processiong function stored in data-raw/costs/AEO/capitalcosts.R, which
uses a gdpdeflator series to get costs in USD$2010. The reported overnight categories are also
mapped to our technology categories. Original reports, manually-compiled datafile, and technology mapping
included in data-raw/costs/AEO/.

\describe {
\item{yr}{Year of reported data}
\item{overnightcategory}{Technology category}
\item{overnight}{Base overnight costs, in $/kW}
\item{om.var}{Variable operations & management costs, in $/kW/yr}
\item{om.fixed}{Fixed operations & management costs, in $/MWh}
}
}
