% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{energy.markets}
\alias{energy.markets}
\alias{energy.markets-package}
\title{energy.markets: A data package for U.S. electricity capacity, generation, and investment.}
\description{
The energy.markets package provides cleaned and processed data for electricity, generation, and investment,
as well as calculations of plant-level capacity factors and levelized cost. The final levelized cost data is
used externally as the target dataset for a multinomial logistic regression, as part of JGCRI hindcasting
experiments.
}
\section{Data sources}{

Capacity and generation data are processed from raw data files providing the responses to EIA forms.

Investment data is processed from Annual Energy Outlook forms published by the EIA.

Fuel price data is processed from tables provided in back issues of the EIA Electric Power Monthly.

A GDP deflator series for the St. Louis Federal Reserve is used to normalize USD$ values to 2010.
}

\section{Calculations}{


Documentation.
}

