% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fuelprices}
\alias{fuelprices}
\title{Fuel Prices}
\format{A data framewith 94 and 3 variables}
\source{
\url{https://www.eia.gov/electricity/monthly/backissues.html}
}
\usage{
data(fuelprices)
}
\description{
Fossil fuel prices taken from EIA Electric Power Monthly reports.
}
\details{
A data file compiles prices reported in the following issues:
June 1996 - Table 26
January 2010 - Table 4.2
May 2016 - Table 4.2

The original reports and compiled data file can be found in data-raw/costs/fuel/.

Uranium source??

\describe{
\item{yr}{Year of reported data}
\item{fuel.general}{Fuel category}
\item{fuel.price}{Fuel price, in $/BTU}
}
}
