% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generation.unmapped}
\alias{generation.unmapped}
\title{Electricity Generation Net Output, Unmapped}
\format{A data framewith 208679 and 7 variables}
\source{
\url{https://www.eia.gov/electricity/data/eia923/}
}
\usage{
data(generation.unmapped)
}
\description{
U.S. electrical generation net output compiled from EIA form 906, 920, and 923 (depending on year), using
native EIA primemover and fuel codes.
}
\details{
#' See data-raw/generation/1990to2000_utilities.R and 2001to2016_utilities.R for functions which
produced the unmapped data set from the raw original data files provided by the EIA.

Generation is negative where generator consumed more electricity than it produced. These rows are filtered out
when calculating capacity factors.

\describe{
\item{yr}{Year of reported data}
\item{utilcode}{EIA-assigned utility code}
\item{plntcode}{EIA-assigned plant code}
\item{primemover}{Native EIA prime mover code}
\item{fuel}{Native EIA fuel code}
\item{generation}{Reported net electrical output, MWh}
\item{consumption}{}
}
}
