% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{levelizedcosts}
\alias{levelizedcosts}
\title{Levelized Capital Costs}
\format{A data framewith 70084 and 11 variables}
\usage{
data(levelizedcosts)
}
\description{
Capital costs with levelized overnight and fixed O&M costs.
}
\details{
Equation for levelized cost can be found in data-raw/costs/levelize.R, which involves
capacity factors dataset and a fixed charge rate, which spreads the cost of the investment over the
expected lifetime of the capacity addition. The fixed charge rate is taken to be 0.13 in order to
match GCAM calibration.

\describe{
\item{yr}{Year of reported data}
\item{startyr}{Average of generation-level startyr, weighted by capacity}
\item{plntcode}{EIA-assigned plant code}
\item{overnightcategory}{Technology category}
\item{fuel.general}{Fuel category}
\item{LCOE_Capital}{Levelized overnight cost, in $/MWh}
\item{LCOE_FOM}{Levelized fixed O&M, in $/MWh}
\item{LCOE_VOM}{Variable O&m, in $/MWh}
\item{LCOE_wo_Fuel}{Capital + FOM + VOM, in $/MWh}
\item{LCOE_Fuel}{Fuel price adjusted to AEO-reported heatrate, in $/MWh}
\item{LCOE}{LCOE_wo_Fuel + LCOE_Fuel, in $/MWh}
}
}
