% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mapping}
\alias{mapping}
\title{Mapping to GCAM Fuel-Tech Options}
\format{A data framewith 181 and 4 variables}
\usage{
data(mapping)
}
\description{
Takes all unique observations of prime mover and fuel, and provides a technology category and
fuel category to describe the data at a courser level than the EIA-native prime mover (33 observed) and fuel
codes (61 observed).
}
\details{
\describe {
\item{primemover}{Native EIA prime mover code}
\item{fuel}{Native EIA fuel code}
\item{overnightcategory}{Technology category}
\item{fuel.general}{Fuel category}
}
}
