% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{marginalcosts}
\alias{marginalcosts}
\title{Marginal Costs}
\format{A data framewith 190 and 4 variables}
\usage{
data(marginalcosts)
}
\description{
Fuel price scaled by fuel efficiency, which is calculated from average heatrates given for a specified
technology and fuel in a given year.
}
\details{
See 'Marginal Costs' section of data-raw/generatedata.R for this calculation. Heatrates included at
data-raw/costs/fuel/avghr.csv.

Source of avghr.csv??

\describe{
\item{yr}{Year of reported data}
\item{overnightcategory}{Technology category}
\item{fuel.general}{Fuel category}
\item{marginal.cost}{Marginal cost of producing electricity using specified technology and fuel, in $/MWh}
}
}
