% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketshare.R
\name{marketshares}
\alias{marketshares}
\title{Calculate electricity-generating capacity market shares}
\usage{
marketshares(generators, mapping, type, additions = FALSE)
}
\arguments{
\item{generators}{Generator-level data holding capacity numbers}

\item{mapping}{Map overnightcategory and fuel.general onto primemover and fuel}

\item{type}{"Fuel" market (top-level) or "tech" market (nested)}

\item{additions}{Logical indicating whether to calculate share of capacity additions for each year}
}
\value{
data.frame with market shares
}
\description{
Calculate electricity-generating capacity market shares
}
