% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{master}
\alias{master}
\title{Master Dataset}
\format{A data framewith 106468 and 10 variables}
\usage{
data(master)
}
\description{
This dataset joins levelizedcosts, marginalcosts, and cap.gen.joined, so that it contains all the
necessary columns for JGCRI's electricity hindcasting project.
}
\details{
\describe {
\item{yr}{Year of reported data}
\item{plntcode}{EIA-assigned plant code}
\item{overnightcategory}{Technology category}
\item{fuel.general}{Fuel category}
\item{overnight.lev}{Levelized overnight cost, in $/MWh}
\item{om.fixed.lev}{Levelized fixed O&M, in $/MWh}
\item{om.var}{Variable O&m, in $/MWh}
\item{marginal.cost}{Marginal cost of producing electricity using specified technology and fuel, in $/MWh}
\item{nameplate}{Reported nameplate capacity, MW}
\item{generation}{Reported net electrical output, MWh}
}
}
