% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summaryCalc}
\alias{summaryCalc}
\title{Calculate summary time-series of master dataset for variable of interest}
\usage{
summaryCalc(master, column, weights = NULL)
}
\arguments{
\item{master}{data.frame, provided in package data}

\item{column}{unquoted column name, c(capacity, generation, lcoe)}
}
\value{
data.frame using keys (oc, fg, yr), reporting yearly averages for column-variable and ultimate averages for missing (oc, fg, yr) rows
}
\description{
Calculate summary time-series of master dataset for variable of interest
}
