% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{capacity.unmapped}
\alias{capacity.unmapped}
\title{Electrical Capacity, Unmapped}
\format{A data framewith 245936 and 7 variables}
\source{
\url{https://www.eia.gov/electricity/data/eia860/}
}
\usage{
data(capacity.unmapped)
}
\description{
U.S. electrical capacity compiled from EIA form 860 for 1990-2016, using native EIA
primemover and fuel codes.
}
\details{
See data-raw/generators/1990to2000_utilities.R and 2001to2016_utilities.R for functions which
produced the unmapped data set from the raw original data files provided by the EIA.

\describe{
\item{yr}{Year of reported data}
\item{utilcode}{EIA-assigned utility code}
\item{plntcode}{EIA-assigned plant code}
\item{primemover}{Native EIA prime mover code}
\item{fuel}{Native EIA fuel code}
\item{vintage}{First year capacity is available}
\item{capacity}{Reported nameplate capacity, MW}
}
}
