% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{capitalcosts}
\alias{capitalcosts}
\title{Capital Costs}
\format{A data framewith 323 and 6 variables}
\source{
\url{https://www.eia.gov/outlooks/aeo/archive.php}
}
\usage{
data(capitalcosts)
}
\description{
Input assumptions to EIA's National Energy Model System, which is used to produce the Annual
Energy Outlook report, for 1996-2015.
}
\details{
Tables from each report appear under the name 'Cost and Performance Characteristics of New
Central Station Electricity Generating Technologies' and were manually compiled into one table.

The processing for this data is contained in an Excel workbook within the data-raw/costs/ directory. The manually
compiled data is contained in the sheet named "AEOData". As this data is based on expectations of construction costs,
there are some spikes in these numbers; "AEOData_$2010_Smoothed" contains the data with these spikes smoothed out and
converted into USD$2010. "Data4GCAM" is the final capital costs dataset reported in this package, with the native AEO
technology categories mapped to our own set of technologies.

\describe{
\item{yr}{Year of reported data}
\item{technology}{Technology label, native to EIA}
\item{overnight}{Base overnight costs, in $/kW}
\item{om.var}{Variable operations & management costs, in $/kW/yr}
\item{om.fixed}{Fixed operations & management costs, in $/MWh}
\item{heatrate}{Heatrate, BTU/kWh}
}
}
