% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{levelizedcosts}
\alias{levelizedcosts}
\title{Levelized Capital Costs}
\format{A data framewith 82036 and 8 variables}
\usage{
data(levelizedcosts)
}
\description{
Capital costs with levelized overnight and fixed O&M costs.
}
\details{
Equation for levelized cost can be found in data-raw/costs/levelize.R, which involves
capacity factors dataset and a fixed charge rate, which spreads the cost of the investment over the
expected lifetime of the capacity addition. The fixed charge rate is taken to be 0.13 in order to
match GCAM calibration.

\describe {
\item{yr}{Year of reported data}
\item{utilcode}{EIA-assigned utility code}
\item{plntcode}{EIA-assigned plant code}
\item{overnightcategory}{Technology category}
\item{fuel.general}{Fuel category}
\item{overnight.lev}{Levelized overnight cost, in $/MWh}
\item{om.fixed.lev}{Levelized fixed O&M, in $/MWh}
\item{om.var}{Variable O&m, in $/MWh}
}
}
